# ahms-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ayurvedic Hospital Management System — Phase 1 API (Auth, Roles, Departments, Patients, Doctors, Appointments, Dashboard).",
        "title": "AHMS Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/patients": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Optional ?search= matches full name, mobile, or UHID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "patients"
                ],
                "summary": "List / search patients",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/patients.PatientResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "If a patient already exists with the same mobile and force=false, responds 409 with the existing matches so the receptionist can confirm before re-submitting with force=true.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "patients"
                ],
                "summary": "Register a new patient (auto-generates UHID)",
                "parameters": [
                    {
                        "description": "Patient",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/patients.CreatePatientRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/patients.PatientResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/patients.DuplicateMobileResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/patients/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "patients"
                ],
                "summary": "Get a patient by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/patients.PatientResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "patients"
                ],
                "summary": "Update a patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Patient",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/patients.UpdatePatientRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/patients.PatientResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "patients"
                ],
                "summary": "Delete (deactivate) a patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "patients.CreatePatientRequest": {
            "type": "object",
            "required": [
                "full_name",
                "gender",
                "mobile"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 500
                },
                "age": {
                    "type": "integer",
                    "maximum": 150,
                    "minimum": 0
                },
                "allergies": {
                    "type": "string"
                },
                "alternate_mobile": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 10
                },
                "blood_group": {
                    "type": "string",
                    "maxLength": 5
                },
                "blood_pressure": {
                    "type": "string",
                    "maxLength": 20
                },
                "branch": {
                    "type": "string",
                    "maxLength": 100
                },
                "chronic_diseases": {
                    "type": "string"
                },
                "city": {
                    "type": "string",
                    "maxLength": 100
                },
                "country": {
                    "type": "string",
                    "maxLength": 100
                },
                "current_medication": {
                    "type": "string"
                },
                "district": {
                    "type": "string",
                    "maxLength": 100
                },
                "dob": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "emergency_contact": {
                    "type": "string",
                    "maxLength": 15
                },
                "emergency_contact_address": {
                    "type": "string",
                    "maxLength": 500
                },
                "emergency_contact_name": {
                    "type": "string",
                    "maxLength": 150
                },
                "emergency_contact_relation": {
                    "type": "string",
                    "maxLength": 50
                },
                "force": {
                    "type": "boolean"
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 2
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "MALE",
                        "FEMALE",
                        "OTHER"
                    ]
                },
                "height_cm": {
                    "type": "number",
                    "maximum": 250,
                    "minimum": 30
                },
                "marital_status": {
                    "type": "string",
                    "maxLength": 20
                },
                "mobile": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 10
                },
                "occupation": {
                    "type": "string",
                    "maxLength": 100
                },
                "photo_url": {
                    "type": "string",
                    "maxLength": 500
                },
                "pincode": {
                    "type": "string",
                    "maxLength": 10
                },
                "pulse": {
                    "type": "string",
                    "maxLength": 20
                },
                "referred_by": {
                    "type": "string",
                    "maxLength": 150
                },
                "registration_type": {
                    "type": "string",
                    "enum": [
                        "WALK_IN",
                        "ONLINE",
                        "REFERRAL"
                    ]
                },
                "remarks": {
                    "type": "string"
                },
                "state": {
                    "type": "string",
                    "maxLength": 100
                },
                "sugar": {
                    "type": "string",
                    "maxLength": 20
                },
                "weight_kg": {
                    "type": "number",
                    "maximum": 300,
                    "minimum": 1
                }
            }
        },
        "patients.DuplicateMobileResponse": {
            "type": "object",
            "properties": {
                "existing_patients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/patients.PatientResponse"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "patients.PatientResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "age": {
                    "type": "integer"
                },
                "allergies": {
                    "type": "string"
                },
                "alternate_mobile": {
                    "type": "string"
                },
                "blood_group": {
                    "type": "string"
                },
                "blood_pressure": {
                    "type": "string"
                },
                "bmi": {
                    "type": "number"
                },
                "branch": {
                    "type": "string"
                },
                "chronic_diseases": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "current_medication": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "dob": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "emergency_contact": {
                    "type": "string"
                },
                "emergency_contact_address": {
                    "type": "string"
                },
                "emergency_contact_name": {
                    "type": "string"
                },
                "emergency_contact_relation": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "height_cm": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "marital_status": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                },
                "occupation": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                },
                "pincode": {
                    "type": "string"
                },
                "pulse": {
                    "type": "string"
                },
                "referred_by": {
                    "type": "string"
                },
                "registration_type": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "sugar": {
                    "type": "string"
                },
                "uhid": {
                    "type": "string"
                },
                "weight_kg": {
                    "type": "number"
                }
            }
        },
        "patients.UpdatePatientRequest": {
            "type": "object",
            "required": [
                "full_name",
                "gender",
                "is_active",
                "mobile"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 500
                },
                "age": {
                    "type": "integer",
                    "maximum": 150,
                    "minimum": 0
                },
                "allergies": {
                    "type": "string"
                },
                "alternate_mobile": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 10
                },
                "blood_group": {
                    "type": "string",
                    "maxLength": 5
                },
                "blood_pressure": {
                    "type": "string",
                    "maxLength": 20
                },
                "branch": {
                    "type": "string",
                    "maxLength": 100
                },
                "chronic_diseases": {
                    "type": "string"
                },
                "city": {
                    "type": "string",
                    "maxLength": 100
                },
                "country": {
                    "type": "string",
                    "maxLength": 100
                },
                "current_medication": {
                    "type": "string"
                },
                "district": {
                    "type": "string",
                    "maxLength": 100
                },
                "dob": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "emergency_contact": {
                    "type": "string",
                    "maxLength": 15
                },
                "emergency_contact_address": {
                    "type": "string",
                    "maxLength": 500
                },
                "emergency_contact_name": {
                    "type": "string",
                    "maxLength": 150
                },
                "emergency_contact_relation": {
                    "type": "string",
                    "maxLength": 50
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 2
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "MALE",
                        "FEMALE",
                        "OTHER"
                    ]
                },
                "height_cm": {
                    "type": "number",
                    "maximum": 250,
                    "minimum": 30
                },
                "is_active": {
                    "type": "boolean"
                },
                "marital_status": {
                    "type": "string",
                    "maxLength": 20
                },
                "mobile": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 10
                },
                "occupation": {
                    "type": "string",
                    "maxLength": 100
                },
                "photo_url": {
                    "type": "string",
                    "maxLength": 500
                },
                "pincode": {
                    "type": "string",
                    "maxLength": 10
                },
                "pulse": {
                    "type": "string",
                    "maxLength": 20
                },
                "referred_by": {
                    "type": "string",
                    "maxLength": 150
                },
                "registration_type": {
                    "type": "string",
                    "enum": [
                        "WALK_IN",
                        "ONLINE",
                        "REFERRAL"
                    ]
                },
                "remarks": {
                    "type": "string"
                },
                "state": {
                    "type": "string",
                    "maxLength": 100
                },
                "sugar": {
                    "type": "string",
                    "maxLength": 20
                },
                "weight_kg": {
                    "type": "number",
                    "maximum": 300,
                    "minimum": 1
                }
            }
        },
...
        "utils.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}